move model loading related tests
